requirements.txt
numpy
scipy
matplotlib
sphinxcontrib-bibtex